+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.17.4
 
 * warnings for some use cases of Closeable
+* added an optional top limit for IDAllocator